+++ src/test/java/com/google/cloud/spanner/pgadapter/JdbcMockServerTest.java
@@ -388,6 +388,44 @@ public class JdbcMockServerTest extends AbstractMockServerTest {
   }
 
   @Test
+  public void testSetWellKnownClient() throws SQLException {
+    for (String client : new String[] {"pgx", "npgsql", "sqlalchemy2"}) {
+      try (Connection connection =
+          DriverManager.getConnection(
+              String.format(
+                  "jdbc:postgresql://localhost:%d/?options=-c%%20spanner.well_known_client=%s",
+                  pgServer.getLocalPort(), client))) {
+        try (ResultSet resultSet =
+            connection.createStatement().executeQuery("show spanner.well_known_client")) {
+          assertTrue(resultSet.next());
+          assertEquals(client.toUpperCase(), resultSet.getString(1));
+          assertFalse(resultSet.next());
+        }
+        PSQLException exception =
+            assertThrows(
+                PSQLException.class,
+                () -> connection.createStatement().execute("set spanner.well_known_client='foo'"));
+        assertNotNull(exception.getServerErrorMessage());
+        assertEquals(
+            "parameter \"spanner.well_known_client\" cannot be set after connection start",
+            exception.getServerErrorMessage().getMessage());
+      }
+    }
+    try (Connection connection =
+        DriverManager.getConnection(
+            String.format(
+                "jdbc:postgresql://localhost:%d/?options=-c%%20spanner.well_known_client=%s",
+                pgServer.getLocalPort(), "foo"))) {
+      try (ResultSet resultSet =
+          connection.createStatement().executeQuery("show spanner.well_known_client")) {
+        assertTrue(resultSet.next());
+        assertEquals("foo", resultSet.getString(1));
+        assertFalse(resultSet.next());
+      }
+    }
+  }
+
+  @Test
   public void testPreparedStatementParameterMetadata() throws SQLException {
     String sql = "SELECT * FROM foo WHERE id=? or value=?";
     String pgSql = "SELECT * FROM foo WHERE id=$1 or value=$2";
@@ -1121,16 +1159,18 @@ public class JdbcMockServerTest extends AbstractMockServerTest {
     mockSpanner.putStatementResult(
         StatementResult.exception(
             Statement.newBuilder(pgSql).bind("p1").to(1L).build(),
-            Status.NOT_FOUND
-                .withDescription("Table non_existing_table not found")
+            Status.INVALID_ARGUMENT
+                .withDescription("relation \"non_existing_table\" does not exist")
                 .asRuntimeException()));
     try (Connection connection = DriverManager.getConnection(createUrl())) {
       try (PreparedStatement preparedStatement = connection.prepareStatement(sql)) {
         preparedStatement.setLong(1, 1L);
-        SQLException exception = assertThrows(SQLException.class, preparedStatement::executeQuery);
+        PSQLException exception =
+            assertThrows(PSQLException.class, preparedStatement::executeQuery);
         assertEquals(
-            "ERROR: Table non_existing_table not found - Statement: 'select * from non_existing_table where id=$1'",
+            "ERROR: relation \"non_existing_table\" does not exist - Statement: 'select * from non_existing_table where id=$1'",
             exception.getMessage());
+        assertEquals(SQLState.UndefinedTable.toString(), exception.getSQLState());
       }
     }
 
@@ -3891,7 +3931,7 @@ public class JdbcMockServerTest extends AbstractMockServerTest {
           }
           count++;
         }
-        assertEquals(360, count);
+        assertEquals(361, count);
       }
     }
   }